requirements.txt
Pillow==9.4.0
webcolors==1.12
